+++ packages/jest-docblock/src/index.js
@@ -13,11 +13,11 @@ import {EOL} from 'os';
 const commentEndRe = /\*\/$/;
 const commentStartRe = /^\/\*\*/;
 const docblockRe = /^\s*(\/\*\*?(.|\r?\n)*?\*\/)/;
-const lineCommentRe = /\/\/([^\r\n]*)/g;
+const lineCommentRe = /(^|\s+)\/\/([^\r\n]*)/g;
 const ltrimRe = /^\s*/;
 const rtrimRe = /\s*$/;
 const ltrimNewlineRe = /^(\r?\n)+/;
-const multilineRe = /(?:^|\r?\n) *(@[^\r\n]*?) *\r?\n *([^@\r\n\s][^@\r\n]+?) *\r?\n/g;
+const multilineRe = /(?:^|\r?\n) *(@[^\r\n]*?) *\r?\n *(?![^@\r\n]*\/\/[^]*)([^@\r\n\s][^@\r\n]+?) *\r?\n/g;
 const propertyRe = /(?:^|\r?\n) *@(\S+) *([^\r\n]*)/g;
 const stringStartRe = /(\r?\n|^) *\* ?/g;
 
@@ -45,7 +45,6 @@ export function parseWithComments(
   docblock = docblock
     .replace(commentStartRe, '')
     .replace(commentEndRe, '')
-    .replace(lineCommentRe, '')
     .replace(stringStartRe, '$1');
 
   // Normalize multi-line directives
@@ -64,7 +63,8 @@ export function parseWithComments(
 
   let match;
   while ((match = propertyRe.exec(docblock))) {
-    result[match[1]] = match[2];
+    // strip linecomments from pragmas
+    result[match[1]] = match[2].replace(lineCommentRe, '');
   }
   return {comments, pragmas: result};
 }